pyproject.toml
[tool.poetry]
name = "crypto-price"
version = "0.1.0"
description = ""
authors = ["Rodrigo Vaccari <rvaccari@gmail.com>"]

[tool.mypy]
strict = true

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | .coverage*
  | manage.py
)
'''

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-extensions = "^3.1.3"
django-ninja = "^0.13.2"
python-decouple = "^3.4"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
arrow = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-django = "^4.4.0"
black = "19.10b0"
pytest-black = "^0.3.12"
pytest-mypy = "^0.8.1"
pytest-xdist = "^2.3.0"
pytest-cov = "^2.12.1"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
